Font-change and Color-change additions
Moved the feature that allowed a font change from a placeholder location, to the settings page. Completed font change implementation.

Proof-of-concept implementation of color-change is completed. Massive re-coding of every color used in the app is needed. The majority of our text/window theming is not centralized. This needs changing for color-changes to be done by the user.